changes to form nonces
If the url page has atleast one nonce: set the nonce field to yes. The
other forms that may have no nonces will be listed in the details.

If the url page has no forms at all: set the nonce field to yes. This is
because we may incorrectly add it to a page that is not fully secure.